requirements.txt
streamlit
pandas==1.5.3
numpy==1.24.4
matplotlib==3.9.4
plotly==6.0.1
scikit-learn==1.6.1
seaborn==0.13.2
pymongo==4.12.1
kafka-python==2.0.2



